refactor(CAG): restructured/ cleaned up CAG a bit * extracted overcutInsideCorner to a seperade module, aliased it inside CAG
