+++ src/CAG.js
@@ -14,6 +14,8 @@ const {fromSides, fromFakeCSG, fromPoints, fromPointsNoCheck} = require('./CAGFa
 const canonicalize = require('./utils/canonicalize')
 const retesselate = require('./utils/retesellate')
 
+const overCutInsideCorners = require('./api/cnc/overcutInsideCorner')
+
 /**
  * Class CAG
  * Holds a solid area geometry like CSG but 2D.
@@ -27,194 +29,6 @@ let CAG = function () {
 }
 
 CAG.prototype = {
-  toString: function () {
-    let result = 'CAG (' + this.sides.length + ' sides):\n'
-    this.sides.map(function (side) {
-      result += '  ' + side.toString() + '\n'
-    })
-    return result
-  },
-
-  _toCSGWall: function (z0, z1) {
-    let polygons = this.sides.map(function (side) {
-      return side.toPolygon3D(z0, z1)
-    })
-    return fromPolygons(polygons)
-  },
-
-  _toVector3DPairs: function (m) {
-        // transform m
-    let pairs = this.sides.map(function (side) {
-      let p0 = side.vertex0.pos
-      let p1 = side.vertex1.pos
-      return [Vector3D.Create(p0.x, p0.y, 0),
-        Vector3D.Create(p1.x, p1.y, 0)]
-    })
-    if (typeof m !== 'undefined') {
-      pairs = pairs.map(function (pair) {
-        return pair.map(function (v) {
-          return v.transform(m)
-        })
-      })
-    }
-    return pairs
-  },
-
-    /*
-     * transform a cag into the polygons of a corresponding 3d plane, positioned per options
-     * Accepts a connector for plane positioning, or optionally
-     * single translation, axisVector, normalVector arguments
-     * (toConnector has precedence over single arguments if provided)
-     */
-  _toPlanePolygons: function (options) {
-    const defaults = {
-      flipped: false
-    }
-    options = Object.assign({}, defaults, options)
-    let {flipped} = options
-    // reference connector for transformation
-    let origin = [0, 0, 0]
-    let defaultAxis = [0, 0, 1]
-    let defaultNormal = [0, 1, 0]
-    let thisConnector = new Connector(origin, defaultAxis, defaultNormal)
-    // translated connector per options
-    let translation = options.translation || origin
-    let axisVector = options.axisVector || defaultAxis
-    let normalVector = options.normalVector || defaultNormal
-    // will override above if options has toConnector
-    let toConnector = options.toConnector ||
-            new Connector(translation, axisVector, normalVector)
-    // resulting transform
-    let m = thisConnector.getTransformationTo(toConnector, false, 0)
-    // create plane as a (partial non-closed) CSG in XY plane
-    let bounds = this.getBounds()
-    bounds[0] = bounds[0].minus(new Vector2D(1, 1))
-    bounds[1] = bounds[1].plus(new Vector2D(1, 1))
-    let csgshell = this._toCSGWall(-1, 1)
-    let csgplane = fromPolygons([new Polygon([
-      new Vertex3D(new Vector3D(bounds[0].x, bounds[0].y, 0)),
-      new Vertex3D(new Vector3D(bounds[1].x, bounds[0].y, 0)),
-      new Vertex3D(new Vector3D(bounds[1].x, bounds[1].y, 0)),
-      new Vertex3D(new Vector3D(bounds[0].x, bounds[1].y, 0))
-    ])])
-    if (flipped) {
-      csgplane = csgplane.invert()
-    }
-    // intersectSub -> prevent premature retesselate/canonicalize
-    csgplane = csgplane.intersectSub(csgshell)
-    // only keep the polygons in the z plane:
-    let polys = csgplane.polygons.filter(function (polygon) {
-      return Math.abs(polygon.plane.normal.z) > 0.99
-    })
-    // finally, position the plane per passed transformations
-    return polys.map(function (poly) {
-      return poly.transform(m)
-    })
-  },
-
-  /*
-    * given 2 connectors, this returns all polygons of a "wall" between 2
-    * copies of this cag, positioned in 3d space as "bottom" and
-    * "top" plane per connectors toConnector1, and toConnector2, respectively
-    */
-  _toWallPolygons: function (options) {
-        // normals are going to be correct as long as toConn2.point - toConn1.point
-        // points into cag normal direction (check in caller)
-        // arguments: options.toConnector1, options.toConnector2, options.cag
-        //     walls go from toConnector1 to toConnector2
-        //     optionally, target cag to point to - cag needs to have same number of sides as this!
-    let origin = [0, 0, 0]
-    let defaultAxis = [0, 0, 1]
-    let defaultNormal = [0, 1, 0]
-    let thisConnector = new Connector(origin, defaultAxis, defaultNormal)
-        // arguments:
-    let toConnector1 = options.toConnector1
-        // let toConnector2 = new Connector([0, 0, -30], defaultAxis, defaultNormal);
-    let toConnector2 = options.toConnector2
-    if (!(toConnector1 instanceof Connector && toConnector2 instanceof Connector)) {
-      throw new Error('could not parse Connector arguments toConnector1 or toConnector2')
-    }
-    if (options.cag) {
-      if (options.cag.sides.length !== this.sides.length) {
-        throw new Error('target cag needs same sides count as start cag')
-      }
-    }
-        // target cag is same as this unless specified
-    let toCag = options.cag || this
-    let m1 = thisConnector.getTransformationTo(toConnector1, false, 0)
-    let m2 = thisConnector.getTransformationTo(toConnector2, false, 0)
-    let vps1 = this._toVector3DPairs(m1)
-    let vps2 = toCag._toVector3DPairs(m2)
-
-    let polygons = []
-    vps1.forEach(function (vp1, i) {
-      polygons.push(new Polygon([
-        new Vertex3D(vps2[i][1]), new Vertex3D(vps2[i][0]), new Vertex3D(vp1[0])]))
-      polygons.push(new Polygon([
-        new Vertex3D(vps2[i][1]), new Vertex3D(vp1[0]), new Vertex3D(vp1[1])]))
-    })
-    return polygons
-  },
-
-    /**
-     * Convert to a list of points.
-     * @return {points[]} list of points in 2D space
-     */
-  toPoints: function () {
-    let points = this.sides.map(function (side) {
-      let v0 = side.vertex0
-      // let v1 = side.vertex1
-      return v0.pos
-    })
-    // due to the logic of fromPoints()
-    // move the first point to the last
-    if (points.length > 0) {
-      points.push(points.shift())
-    }
-    return points
-  },
-
-    /** Convert to compact binary form.
-   * See fromCompactBinary.
-   * @return {CompactBinary}
-   */
-  toCompactBinary: function () {
-    let cag = this.canonicalized()
-    let numsides = cag.sides.length
-    let vertexmap = {}
-    let vertices = []
-    let numvertices = 0
-    let sideVertexIndices = new Uint32Array(2 * numsides)
-    let sidevertexindicesindex = 0
-    cag.sides.map(function (side) {
-      [side.vertex0, side.vertex1].map(function (v) {
-        let vertextag = v.getTag()
-        let vertexindex
-        if (!(vertextag in vertexmap)) {
-          vertexindex = numvertices++
-          vertexmap[vertextag] = vertexindex
-          vertices.push(v)
-        } else {
-          vertexindex = vertexmap[vertextag]
-        }
-        sideVertexIndices[sidevertexindicesindex++] = vertexindex
-      })
-    })
-    let vertexData = new Float64Array(numvertices * 2)
-    let verticesArrayIndex = 0
-    vertices.map(function (v) {
-      let pos = v.pos
-      vertexData[verticesArrayIndex++] = pos._x
-      vertexData[verticesArrayIndex++] = pos._y
-    })
-    let result = {
-      'class': 'CAG',
-      sideVertexIndices: sideVertexIndices,
-      vertexData: vertexData
-    }
-    return result
-  },
-
   union: function (cag) {
     let cags
     if (cag instanceof Array) {
@@ -278,9 +92,17 @@ CAG.prototype = {
     return result
   },
 
-    // see http://local.wasp.uwa.edu.au/~pbourke/geometry/polyarea/ :
-    // Area of the polygon. For a counter clockwise rotating polygon the area is positive, otherwise negative
-    // Note(bebbi): this looks wrong. See polygon getArea()
+  flipped: function () {
+    let newsides = this.sides.map(function (side) {
+      return side.flipped()
+    })
+    newsides.reverse()
+    return fromSides(newsides)
+  },
+
+  // see http://local.wasp.uwa.edu.au/~pbourke/geometry/polyarea/ :
+  // Area of the polygon. For a counter clockwise rotating polygon the area is positive, otherwise negative
+  // Note(bebbi): this looks wrong. See polygon getArea()
   area: function () {
     let polygonArea = 0
     this.sides.map(function (side) {
@@ -288,14 +110,6 @@ CAG.prototype = {
     })
     polygonArea *= 0.5
     return polygonArea
-  },
-
-  flipped: function () {
-    let newsides = this.sides.map(function (side) {
-      return side.flipped()
-    })
-    newsides.reverse()
-    return fromSides(newsides)
   },
 
   getBounds: function () {
@@ -668,79 +482,196 @@ CAG.prototype = {
     return paths
   },
 
-    /*
-    cag = cag.overCutInsideCorners(cutterradius);
-
-    Using a CNC router it's impossible to cut out a true sharp inside corner. The inside corner
-    will be rounded due to the radius of the cutter. This function compensates for this by creating
-    an extra cutout at each inner corner so that the actual cut out shape will be at least as large
-    as needed.
-    */
-  overCutInsideCorners: function (cutterradius) {
-    let cag = this.canonicalized()
-        // for each vertex determine the 'incoming' side and 'outgoing' side:
-    let pointmap = {} // tag => {pos: coord, from: [], to: []}
-    cag.sides.map(function (side) {
-      if (!(side.vertex0.getTag() in pointmap)) {
-        pointmap[side.vertex0.getTag()] = {
-          pos: side.vertex0.pos,
-          from: [],
-          to: []
-        }
-      }
-      pointmap[side.vertex0.getTag()].to.push(side.vertex1.pos)
-      if (!(side.vertex1.getTag() in pointmap)) {
-        pointmap[side.vertex1.getTag()] = {
-          pos: side.vertex1.pos,
-          from: [],
-          to: []
-        }
-      }
-      pointmap[side.vertex1.getTag()].from.push(side.vertex0.pos)
+  toString: function () {
+    let result = 'CAG (' + this.sides.length + ' sides):\n'
+    this.sides.map(function (side) {
+      result += '  ' + side.toString() + '\n'
     })
-        // overcut all sharp corners:
-    let cutouts = []
-    for (let pointtag in pointmap) {
-      let pointobj = pointmap[pointtag]
-      if ((pointobj.from.length === 1) && (pointobj.to.length === 1)) {
-                // ok, 1 incoming side and 1 outgoing side:
-        let fromcoord = pointobj.from[0]
-        let pointcoord = pointobj.pos
-        let tocoord = pointobj.to[0]
-        let v1 = pointcoord.minus(fromcoord).unit()
-        let v2 = tocoord.minus(pointcoord).unit()
-        let crossproduct = v1.cross(v2)
-        let isInnerCorner = (crossproduct < 0.001)
-        if (isInnerCorner) {
-                    // yes it's a sharp corner:
-          let alpha = v2.angleRadians() - v1.angleRadians() + Math.PI
-          if (alpha < 0) {
-            alpha += 2 * Math.PI
-          } else if (alpha >= 2 * Math.PI) {
-            alpha -= 2 * Math.PI
-          }
-          let midvector = v2.minus(v1).unit()
-          let circlesegmentangle = 30 / 180 * Math.PI // resolution of the circle: segments of 30 degrees
-                    // we need to increase the radius slightly so that our imperfect circle will contain a perfect circle of cutterradius
-          let radiuscorrected = cutterradius / Math.cos(circlesegmentangle / 2)
-          let circlecenter = pointcoord.plus(midvector.times(radiuscorrected))
-                    // we don't need to create a full circle; a pie is enough. Find the angles for the pie:
-          let startangle = alpha + midvector.angleRadians()
-          let deltaangle = 2 * (Math.PI - alpha)
-          let numsteps = 2 * Math.ceil(deltaangle / circlesegmentangle / 2) // should be even
-                    // build the pie:
-          let points = [circlecenter]
-          for (let i = 0; i <= numsteps; i++) {
-            let angle = startangle + i / numsteps * deltaangle
-            let p = Vector2D.fromAngleRadians(angle).times(radiuscorrected).plus(circlecenter)
-            points.push(p)
-          }
-          cutouts.push(fromPoints(points))
-        }
+    return result
+  },
+
+  _toCSGWall: function (z0, z1) {
+    let polygons = this.sides.map(function (side) {
+      return side.toPolygon3D(z0, z1)
+    })
+    return fromPolygons(polygons)
+  },
+
+  _toVector3DPairs: function (m) {
+        // transform m
+    let pairs = this.sides.map(function (side) {
+      let p0 = side.vertex0.pos
+      let p1 = side.vertex1.pos
+      return [Vector3D.Create(p0.x, p0.y, 0),
+        Vector3D.Create(p1.x, p1.y, 0)]
+    })
+    if (typeof m !== 'undefined') {
+      pairs = pairs.map(function (pair) {
+        return pair.map(function (v) {
+          return v.transform(m)
+        })
+      })
+    }
+    return pairs
+  },
+
+  /*
+    * transform a cag into the polygons of a corresponding 3d plane, positioned per options
+    * Accepts a connector for plane positioning, or optionally
+    * single translation, axisVector, normalVector arguments
+    * (toConnector has precedence over single arguments if provided)
+    */
+  _toPlanePolygons: function (options) {
+    const defaults = {
+      flipped: false
+    }
+    options = Object.assign({}, defaults, options)
+    let {flipped} = options
+    // reference connector for transformation
+    let origin = [0, 0, 0]
+    let defaultAxis = [0, 0, 1]
+    let defaultNormal = [0, 1, 0]
+    let thisConnector = new Connector(origin, defaultAxis, defaultNormal)
+    // translated connector per options
+    let translation = options.translation || origin
+    let axisVector = options.axisVector || defaultAxis
+    let normalVector = options.normalVector || defaultNormal
+    // will override above if options has toConnector
+    let toConnector = options.toConnector ||
+            new Connector(translation, axisVector, normalVector)
+    // resulting transform
+    let m = thisConnector.getTransformationTo(toConnector, false, 0)
+    // create plane as a (partial non-closed) CSG in XY plane
+    let bounds = this.getBounds()
+    bounds[0] = bounds[0].minus(new Vector2D(1, 1))
+    bounds[1] = bounds[1].plus(new Vector2D(1, 1))
+    let csgshell = this._toCSGWall(-1, 1)
+    let csgplane = fromPolygons([new Polygon([
+      new Vertex3D(new Vector3D(bounds[0].x, bounds[0].y, 0)),
+      new Vertex3D(new Vector3D(bounds[1].x, bounds[0].y, 0)),
+      new Vertex3D(new Vector3D(bounds[1].x, bounds[1].y, 0)),
+      new Vertex3D(new Vector3D(bounds[0].x, bounds[1].y, 0))
+    ])])
+    if (flipped) {
+      csgplane = csgplane.invert()
+    }
+    // intersectSub -> prevent premature retesselate/canonicalize
+    csgplane = csgplane.intersectSub(csgshell)
+    // only keep the polygons in the z plane:
+    let polys = csgplane.polygons.filter(function (polygon) {
+      return Math.abs(polygon.plane.normal.z) > 0.99
+    })
+    // finally, position the plane per passed transformations
+    return polys.map(function (poly) {
+      return poly.transform(m)
+    })
+  },
+
+  /*
+    * given 2 connectors, this returns all polygons of a "wall" between 2
+    * copies of this cag, positioned in 3d space as "bottom" and
+    * "top" plane per connectors toConnector1, and toConnector2, respectively
+    */
+  _toWallPolygons: function (options) {
+        // normals are going to be correct as long as toConn2.point - toConn1.point
+        // points into cag normal direction (check in caller)
+        // arguments: options.toConnector1, options.toConnector2, options.cag
+        //     walls go from toConnector1 to toConnector2
+        //     optionally, target cag to point to - cag needs to have same number of sides as this!
+    let origin = [0, 0, 0]
+    let defaultAxis = [0, 0, 1]
+    let defaultNormal = [0, 1, 0]
+    let thisConnector = new Connector(origin, defaultAxis, defaultNormal)
+        // arguments:
+    let toConnector1 = options.toConnector1
+        // let toConnector2 = new Connector([0, 0, -30], defaultAxis, defaultNormal);
+    let toConnector2 = options.toConnector2
+    if (!(toConnector1 instanceof Connector && toConnector2 instanceof Connector)) {
+      throw new Error('could not parse Connector arguments toConnector1 or toConnector2')
+    }
+    if (options.cag) {
+      if (options.cag.sides.length !== this.sides.length) {
+        throw new Error('target cag needs same sides count as start cag')
       }
     }
-    let result = cag.subtract(cutouts)
+        // target cag is same as this unless specified
+    let toCag = options.cag || this
+    let m1 = thisConnector.getTransformationTo(toConnector1, false, 0)
+    let m2 = thisConnector.getTransformationTo(toConnector2, false, 0)
+    let vps1 = this._toVector3DPairs(m1)
+    let vps2 = toCag._toVector3DPairs(m2)
+
+    let polygons = []
+    vps1.forEach(function (vp1, i) {
+      polygons.push(new Polygon([
+        new Vertex3D(vps2[i][1]), new Vertex3D(vps2[i][0]), new Vertex3D(vp1[0])]))
+      polygons.push(new Polygon([
+        new Vertex3D(vps2[i][1]), new Vertex3D(vp1[0]), new Vertex3D(vp1[1])]))
+    })
+    return polygons
+  },
+
+    /**
+     * Convert to a list of points.
+     * @return {points[]} list of points in 2D space
+     */
+  toPoints: function () {
+    let points = this.sides.map(function (side) {
+      let v0 = side.vertex0
+      // let v1 = side.vertex1
+      return v0.pos
+    })
+    // due to the logic of fromPoints()
+    // move the first point to the last
+    if (points.length > 0) {
+      points.push(points.shift())
+    }
+    return points
+  },
+
+    /** Convert to compact binary form.
+   * See fromCompactBinary.
+   * @return {CompactBinary}
+   */
+  toCompactBinary: function () {
+    let cag = this.canonicalized()
+    let numsides = cag.sides.length
+    let vertexmap = {}
+    let vertices = []
+    let numvertices = 0
+    let sideVertexIndices = new Uint32Array(2 * numsides)
+    let sidevertexindicesindex = 0
+    cag.sides.map(function (side) {
+      [side.vertex0, side.vertex1].map(function (v) {
+        let vertextag = v.getTag()
+        let vertexindex
+        if (!(vertextag in vertexmap)) {
+          vertexindex = numvertices++
+          vertexmap[vertextag] = vertexindex
+          vertices.push(v)
+        } else {
+          vertexindex = vertexmap[vertextag]
+        }
+        sideVertexIndices[sidevertexindicesindex++] = vertexindex
+      })
+    })
+    let vertexData = new Float64Array(numvertices * 2)
+    let verticesArrayIndex = 0
+    vertices.map(function (v) {
+      let pos = v.pos
+      vertexData[verticesArrayIndex++] = pos._x
+      vertexData[verticesArrayIndex++] = pos._y
+    })
+    let result = {
+      'class': 'CAG',
+      sideVertexIndices: sideVertexIndices,
+      vertexData: vertexData
+    }
     return result
+  },
+
+  overCutInsideCorners: function (cutterradius) {
+    return overCutInsideCorners(this, cutterradius)
   }
 }
